Fix plugin initialization spacing
