+++ frontend/src/components/SettingsPanel.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Box, Typography } from '@mui/material';
+
+export default function SettingsPanel() {
+  return (
+    <Box sx={{ p: 2 }}>
+      <Typography variant="h6">Réglages</Typography>
+    </Box>
+  );
+}
